% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.R
\name{RunPopModel}
\alias{RunPopModel}
\title{Run the RCPP code with specified number of years and simulations}
\usage{
RunPopModel(
  NSIM,
  BurnInPop,
  MaxYear,
  CatchAge,
  MaxAge,
  MaxCell,
  NatMort,
  BHa,
  BHb,
  PF,
  AdultMove,
  Mature,
  Weight,
  Settlement,
  Selectivity,
  FishingMort
)
}
\arguments{
\item{NSIM}{Number of simulations you want to run (Integer)}

\item{BurnInPop}{Population you end up with at the end of your burn in period (3D array, columns are months, rows are cells, layers are ages)}

\item{MaxYear}{The total number of years you want your model to run i.e. 1950-2000 would be 51 years (Integer)}

\item{CatchAge}{Integer age at which fish become legal size for catching, if no MLL set to 1 (Integer)}

\item{MaxAge}{How old the species lives for (Integer)}

\item{MaxCell}{The total number of cells in the model (Integer)}

\item{NatMort}{The natural mortality per time step for your species (Integer)}

\item{BHa}{Beverton-Holt parameter alpha (Double)}

\item{BHb}{Beverton-Holt paramter beta (Double)}

\item{PF}{Proportion of females in the population i.e. 0.5 means 50:50 males and females - this is important for calculating spawning biomass (Double)}

\item{AdultMove}{Probability of a fish moving from one cell to any other cell in the model in one time step (Matrix - rows and columns represent cells so should be same as MaxCell)}

\item{Mature}{Probability that a fish of a given age (in years) in mature in each month of the year (Matrix - rows are fish age in years, columns months)}

\item{Weight}{Expected weight of a fish of a given age (in years) in a given month (Matrix - rows are fish age in years, columns are months)}

\item{Settlement}{Probability recruit will settle in a specific cell in the model (Vector - total number of values should be same as total number of cells)}

\item{Selectivity}{Probability of a fish of a given age (in years) being caught and retained in a given month in a given year, this needs to have the same number of years as the model even if selectivity stays the same every year (3D array - rows are fish age in years, columns are months, layers are years)}

\item{FishingMort}{Expected fishing mortality in each cell of the model, in each month, in each year (3D array - rows are cells, columns are months, layers are years of the model)}
}
\value{
list object with all outputs of model including ages per cell, numbers per cell, catch
}
\description{
Run the RCPP code with specified number of years and simulations
}
